+++ src/main/java/com/liceu/sromerom/controllers/ShareController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 
 import javax.servlet.http.HttpSession;
+import java.util.Arrays;
 
 @Controller
 public class ShareController {
@@ -30,51 +31,45 @@ public class ShareController {
         //Si hi ha parametre en la url, procedirem a enviar el usuaris que amb els que ha compartir i carregarem el select amb tots els usuaris
         if (noteid != null) {
             Long userid = (Long) session.getAttribute("userid");
-            model.addAttribute("username", userService.getUserById(userid).getUsername());
             model.addAttribute("noteid", noteid);
+            model.addAttribute("userNoteOwner", noteService.getNoteById(noteid).getUser().getUsername());
             model.addAttribute("users", userService.getUnsharedUsers(userid, noteid));
             model.addAttribute("usersShared", noteService.getPermissionFromSharedUsers(noteid));
 
 
             if (noteService.isNoteOwner(userid, noteid) || noteService.hasWritePermission(userid, noteid)) {
-                model.addAttribute("owner", noteService.isNoteOwner(userid, noteid));
                 model.addAttribute("action", "/share");
             } else {
-                //REDIRECT A RESTRICTED AREA
                 throw new CustomGenericException("Note permission", "Sorry. You don't have access to this note!");
             }
         } else {
-            //REDIRECT AL HOME
-            return "redirect:/home";
+            throw new CustomGenericException("Note not found", "Sorry. There was a problem trying to get the note");
         }
         return "shares";
     }
 
     @PostMapping("/share")
-    public String postShare(@RequestParam("users[]") String[] toShare, @RequestParam(required = false) String permissionMode, @RequestParam("noteid") Long noteid, Model model) {
+    public String postShare(@RequestParam("users[]") String[] toShare, @RequestParam String permissionMode, @RequestParam("noteid") Long noteid, Model model) {
         Long userid = (Long) session.getAttribute("userid");
         boolean noError = false;
+        model.addAttribute("action", "/share");
 
-        if (toShare != null && noteid != null && toShare.length > 0 && !userService.existsUserShare(noteid, userid, toShare)) {
-            //noError = ns.shareNote(userid, noteid, sharedUsers);
-            System.out.println("-------------------------------------------------------------------------------------------------------");
-            System.out.println("Antesssssssssssssssssssssssss: " + permissionMode);
-            if (permissionMode == null) {
-                System.out.println("Es nullllllllllllllllllllllllllllllll");
-                permissionMode = "readmode";
+        //Procedirem a compartir la nota, si entre els usuaris no hi estan compartits
+        if (toShare != null && toShare.length > 0 && noteid != null && !userService.existsUserShare(noteid, userid, toShare)) {
+            if (noteService.isNoteOwner(userid, noteid) || noteService.hasWritePermission(userid, noteid)) {
+                noError = noteService.shareNote(userid, noteid, permissionMode, toShare);
+            } else {
+                throw new CustomGenericException("Note permission", "Sorry. You don't have access to share this note!");
             }
-
-            System.out.println("Despueeeeeeeeeeeeeeeeeeeeeees: " + permissionMode);
-            noError = noteService.shareNote(userid, noteid, permissionMode, toShare);
         }
 
         if (noError) {
-            //REDIRECT AL HOME PORQUE HA IDO BIEN
-            return "redirect:/home";
+            boolean isSessionUser = Arrays.stream(toShare).anyMatch(username -> username.equalsIgnoreCase(userService.getUserById(userid).getUsername()));
+            if (isSessionUser) return "redirect:/home";
+            return "redirect:/share?id=" + noteid;
         }
 
-        model.addAttribute("noError", false);
-        model.addAttribute("action", "/share");
+        model.addAttribute("noerror", false);
 
         return "shares";
     }
@@ -91,12 +86,10 @@ public class ShareController {
             if (noteService.isNoteOwner(userid, noteid) || noteService.hasWritePermission(userid, noteid)) {
                 model.addAttribute("action", "/deleteShare");
             } else {
-                //REDIRECT A RESTRICTED AREA
                 throw new CustomGenericException("Note permission", "Sorry. You don't have access to this note!");
             }
         } else {
-            //REDIRECT AL HOME
-            return "redirect:/home";
+            throw new CustomGenericException("Note not found", "Sorry. There was a problem trying to get the note");
         }
 
         return "shares";
@@ -105,23 +98,45 @@ public class ShareController {
     @PostMapping("/deleteShare")
     public String postDeleteShare(@RequestParam("users[]") String[] toDeleteShare, @RequestParam("noteid") Long noteid, Model model) {
         Long userid = (Long) session.getAttribute("userid");
-        boolean noError = false;
-        if (noteid != null && toDeleteShare.length > 0 && noteService.isNoteOwner(userid, noteid) || noteService.isSharedNote(userid, noteid) || noteService.hasWritePermission(userid, noteid)) {
-            System.out.println("Es note owner y el ha compartido su nota!!!!!!!!!!!");
-            //Eliminarem sempre i quan existeixi usuaris a eliminar el share i que existeixi un share amb aquells usuaris
-            //Mirar metodo existsUserShare
-            if (toDeleteShare != null && userService.existsUserShare(noteid, userid, toDeleteShare)) {
-                System.out.println("Existe share con estos usuarios!!!!!!!!!!!!!!");
+        boolean noError;
+        model.addAttribute("action", "/deleteShare");
+
+        //Eliminarem sempre i quan existeixi usuaris a eliminar el share i que existeixi un share amb aquells usuaris
+        if (noteid != null && toDeleteShare != null && toDeleteShare.length > 0 && userService.existsUserShare(noteid, userid, toDeleteShare)) {
+            if (noteService.isNoteOwner(userid, noteid) || noteService.isSharedNote(userid, noteid) || noteService.hasWritePermission(userid, noteid)) {
                 noError = noteService.deleteShareNote(userid, noteid, toDeleteShare);
+            } else {
+                throw new CustomGenericException("Note permission", "Sorry. You don't have access to this note!");
             }
 
             if (noError) {
-                return "redirect:/home";
-                //REDIRECT AL HOME...
+                boolean isSessionUser = Arrays.stream(toDeleteShare).anyMatch(username -> username.equalsIgnoreCase(userService.getUserById(userid).getUsername()));
+                if (isSessionUser) return "redirect:/home";
+                return "redirect:/deleteShare?id=" + noteid;
             }
         } else {
-            throw new CustomGenericException("Note permission", "Sorry. You don't have access to this note!");
-            //REDIRECT A RESTRICTED AREA
+            throw new CustomGenericException("Note not found", "Sorry. There was a problem trying to get the note");
+        }
+
+        model.addAttribute("noerror", false);
+        return "shares";
+    }
+
+    @PostMapping("/deleteAllShare")
+    public String deleteAllShare(@RequestParam Long noteid, Model model) {
+        Long userid = (Long) session.getAttribute("userid");
+        boolean noError = false;
+        if (noteid != null && userid != null) {
+            if (!noteService.isSharedNote(userid, noteid) && !noteService.isNoteOwner(userid, noteid) && !noteService.hasWritePermission(userid, noteid)) {
+                throw new CustomGenericException("Note permission", "Sorry. You don't have access to this note!");
+            }
+
+            noError = noteService.deleteAllShareNote(userid, noteid);
+
+            if (noError) {
+
+                return "redirect:/home";
+            }
         }
 
         model.addAttribute("noerror", false);
@@ -129,38 +144,25 @@ public class ShareController {
         return "shares";
     }
 
-    @PostMapping("/deleteAllShare")
-    public String deleteAllShare(@RequestParam Long noteid) {
-        Long userid = (Long) session.getAttribute("userid");
-        if (noteid != null && userid != null) {
-            if (!noteService.isSharedNote(userid, noteid) && !noteService.isNoteOwner(userid, noteid)) {
-                throw new CustomGenericException("Note permission", "Sorry. You don't have access to this note!");
-            }
-
-            boolean noerror = noteService.deleteAllShareNote(userid, noteid);
-        }
-        return "redirect:/home";
-    }
-
     @PostMapping("/updatePermission")
     public String updatePermission(@RequestParam Long noteid, @RequestParam Long shareduserid, @RequestParam String permissionMode, Model model) {
         Long userid = (Long) session.getAttribute("userid");
-        System.out.println("######################################################################");
-        System.out.println(noteid);
-        System.out.println(shareduserid);
-        System.out.println(permissionMode);
         boolean noError = true;
         if (noteid != null && userid != null) {
+            if (!noteService.isNoteOwner(userid, noteid) && !noteService.hasWritePermission(userid, noteid)) {
+                throw new CustomGenericException("Note permission", "Sorry. You don't have access to this note!");
+            }
             noError = noteService.updatePermissionMode(userid, shareduserid, noteid, permissionMode);
         }
 
         if (noError) {
-            //REDIRECT AL HOME PORQUE HA IDO BIEN
-            //return "redirect:/share?id=" + noteid;
-            return "redirect:/home";
+            //boolean isSessionUser = Arrays.stream(toDeleteShare).anyMatch(username -> username.equalsIgnoreCase(userService.getUserById(userid).getUsername()));
+            if (shareduserid.equals(userid)) return "redirect:/home";
+            return "redirect:/share?id=" + noteid;
         }
 
-        model.addAttribute("noError", false);
+        model.addAttribute("noerror", false);
+        model.addAttribute("action", "/updatePermission");
         return "shares";
     }
 }
